// src/types/HTMLAttributes.ts
////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////

import { CSSVariables } from "./CSSVariables"

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////

export type HTMLAttributes = {
  reference?: React.RefObject<HTMLDivElement>
  children?: React.ReactNode
  classes?: string[]
  styles?: CSSVariables
  tabIndex?: number
  isFocusable?: boolean
}

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////